requirements.txt
Flask==3.0.0
openai==1.12.0
beautifulsoup4==4.12.3
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
Jinja2==3.1.4
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
soupsieve==2.6
tqdm==4.67.1